Add support for the X-Sendfile header/module to offload downloads.
With the appropriate config option enabled, Jirafeau will send the
X-Sendfile header instead of sending the file itself. As a result,
Apache/lighttpd will send the file without the intervention of Jirafeau,
and download resumes/seeking will also be enabled for this download.

If the user requested an encrypted file then this will not work, but it
should work in all other cases.